Add files via upload
Neo4J Graph Description

The provided graph is a visual display of the airports in the United States and the connections between them (“HAS_ROUTE”). From the graph we can see that airports that have more routes/connections available are located in the middle of the graph. This is expected as they would be more trafficked. More airports have routes to them; hence their central location in the graph.